Add custom event jobCreated
Have added a custom event to be triggered when a job is successfully created, passing the id and link to the job. This will ease integration for customers as subscribing to this removes the need to poll for new jobs created in the background, and just check for success on each created job.